# Optimize and inline bitboard.Grow