This results in the following microbenchmarks on my Early 2013 2.7 GHz
Mac Book Pro using Go 1.6.2:

```
name            old time/op    new time/op    delta
EmptyHasRoad-8    20.7ns ± 5%    20.4ns ± 4%     ~            (p=0.354 n=10+9)
FullHasRoad-8      197ns ± 7%     162ns ± 6%  -18.04%        (p=0.000 n=10+10)
HasRoadWindy-8    59.4ns ± 4%    53.4ns ± 6%  -10.11%        (p=0.000 n=10+10)
```