@zenazn 2016-05-22
Ah, bingo: this doesn't work on 8x8 boards, and the diagram I drew myself was smaller than that