@nelhage 2016-05-22
I think it also doesn't work on smaller boards; `L` and `R` here prevent wraparound when shifting squares at the edge of the board.