@zenazn 2016-05-22
I also misinterpreted the bitboard as embedding a NxN board in the top corner of an 8x8, which is also not what you're doing. So many things were wrong with my mental model of the universe :)